ai_roofing_archive/20250905T192517Z/find_distributors_wrapper/requirements.txt
fastapi
uvicorn
httpx
google-auth

requests
